remove old AssetsDetectionController and tests
