+++ src/assets/CollectibleDetectionController.test.ts
@@ -32,6 +32,13 @@ describe('CollectibleDetectionController', () => {
       getCollectibleTokenURI: assetsContract.getCollectibleTokenURI.bind(
         assetsContract,
       ),
+      getOwnerOf: assetsContract.getOwnerOf.bind(assetsContract),
+      balanceOfERC1155Collectible: assetsContract.balanceOfERC1155Collectible.bind(
+        assetsContract,
+      ),
+      uriERC1155Collectible: assetsContract.uriERC1155Collectible.bind(
+        assetsContract,
+      ),
     });
 
     collectibleDetection = new CollectibleDetectionController({
@@ -53,6 +60,11 @@ describe('CollectibleDetectionController', () => {
           {
             asset_contract: {
               address: '0x1d963688fe2209a98db35c67a041524822cf04ff',
+              schema_name: 'ERC721',
+            },
+            collection: {
+              name: 'Collection 2577',
+              image_url: 'url',
             },
             description: 'Description 2577',
             image_original_url: 'image/2577.png',
@@ -102,6 +114,11 @@ describe('CollectibleDetectionController', () => {
           {
             asset_contract: {
               address: '0xCE7ec4B2DfB30eB6c0BB5656D33aAd6BFb4001Fc',
+              schema_name: 'ERC721',
+            },
+            collection: {
+              name: 'Collection 2577',
+              image_url: 'url',
             },
             description: 'Description 2577',
             image_url: 'image/2577.png',
@@ -111,6 +128,11 @@ describe('CollectibleDetectionController', () => {
           {
             asset_contract: {
               address: '0x0B0fa4fF58D28A88d63235bd0756EDca69e49e6d',
+              schema_name: 'ERC721',
+            },
+            collection: {
+              name: 'Collection 2577',
+              image_url: 'url',
             },
             description: 'Description 2578',
             image_url: 'image/2578.png',
@@ -120,6 +142,11 @@ describe('CollectibleDetectionController', () => {
           {
             asset_contract: {
               address: '0xebE4e5E773AFD2bAc25De0cFafa084CFb3cBf1eD',
+              schema_name: 'ERC721',
+            },
+            collection: {
+              name: 'Collection 2574',
+              image_url: 'url',
             },
             description: 'Description 2574',
             image_url: 'image/2574.png',
@@ -139,6 +166,11 @@ describe('CollectibleDetectionController', () => {
           {
             asset_contract: {
               address: '0xebE4e5E773AFD2bAc25De0cFafa084CFb3cBf1eD',
+              schema_name: 'ERC721',
+            },
+            collection: {
+              name: 'Collection 2574',
+              image_url: 'url',
             },
             description: 'Description 2574',
             image_url: 'image/2574.png',
@@ -242,7 +274,10 @@ describe('CollectibleDetectionController', () => {
         description: 'Description 2574',
         image: 'image/2574.png',
         name: 'ID 2574',
-        tokenId: 2574,
+        tokenId: '2574',
+        standard: 'ERC721',
+        collectionImage: 'url',
+        collectionName: 'Collection 2574',
       },
     ]);
   });
@@ -254,11 +289,12 @@ describe('CollectibleDetectionController', () => {
     });
     await collectiblesController.addCollectible(
       '0xebE4e5E773AFD2bAc25De0cFafa084CFb3cBf1eD',
-      2573,
+      '2573',
       {
         description: 'Description 2573',
         image: 'image/2573.png',
         name: 'ID 2573',
+        standard: 'ERC721',
       },
     );
     await collectibleDetection.detectCollectibles();
@@ -268,14 +304,18 @@ describe('CollectibleDetectionController', () => {
         description: 'Description 2573',
         image: 'image/2573.png',
         name: 'ID 2573',
-        tokenId: 2573,
+        standard: 'ERC721',
+        tokenId: '2573',
       },
       {
         address: '0xebE4e5E773AFD2bAc25De0cFafa084CFb3cBf1eD',
         description: 'Description 2574',
         image: 'image/2574.png',
         name: 'ID 2574',
-        tokenId: 2574,
+        tokenId: '2574',
+        standard: 'ERC721',
+        collectionImage: 'url',
+        collectionName: 'Collection 2574',
       },
     ]);
   });
@@ -290,7 +330,7 @@ describe('CollectibleDetectionController', () => {
     expect(collectiblesController.state.ignoredCollectibles).toHaveLength(0);
     collectiblesController.removeAndIgnoreCollectible(
       '0x1d963688fe2209a98db35c67a041524822cf04ff',
-      2577,
+      '2577',
     );
     await collectibleDetection.detectCollectibles();
     expect(collectiblesController.state.collectibles).toHaveLength(0);
@@ -323,21 +363,30 @@ describe('CollectibleDetectionController', () => {
       description: 'Description 2574',
       image: 'image/2574.png',
       name: 'ID 2574',
-      tokenId: 2574,
+      tokenId: '2574',
+      standard: 'ERC721',
+      collectionImage: 'url',
+      collectionName: 'Collection 2574',
     };
     const collectibleGG2574 = {
       address: '0xCE7ec4B2DfB30eB6c0BB5656D33aAd6BFb4001Fc',
       description: 'Description 2574',
       image: 'image/2574.png',
       name: 'ID 2574',
-      tokenId: 2574,
+      tokenId: '2574',
+      standard: 'ERC721',
+      collectionImage: 'url',
+      collectionName: 'Collection 2574',
     };
     const collectibleII2577 = {
       address: '0x0B0fa4fF58D28A88d63235bd0756EDca69e49e6d',
       description: 'Description 2577',
       image: 'image/2577.png',
       name: 'ID 2577',
-      tokenId: 2577,
+      tokenId: '2577',
+      standard: 'ERC721',
+      collectionImage: 'url',
+      collectionName: 'Collection 2577',
     };
     const collectibleContractHH = {
       address: '0xebE4e5E773AFD2bAc25De0cFafa084CFb3cBf1eD',
@@ -405,6 +454,11 @@ describe('CollectibleDetectionController', () => {
           {
             asset_contract: {
               address: '0x0B0fa4fF58D28A88d63235bd0756EDca69e49e6d',
+              schema_name: 'ERC721',
+            },
+            collection: {
+              name: 'Collection 2577',
+              image_url: 'url',
             },
             description: 'Description 2577',
             image_url: 'image/2577.png',
@@ -414,6 +468,11 @@ describe('CollectibleDetectionController', () => {
           {
             asset_contract: {
               address: '0xCE7ec4B2DfB30eB6c0BB5656D33aAd6BFb4001Fc',
+              schema_name: 'ERC721',
+            },
+            collection: {
+              name: 'Collection 2574',
+              image_url: 'url',
             },
             description: 'Description 2574',
             image_url: 'image/2574.png',
@@ -423,6 +482,11 @@ describe('CollectibleDetectionController', () => {
           {
             asset_contract: {
               address: '0xebE4e5E773AFD2bAc25De0cFafa084CFb3cBf1eD',
+              schema_name: 'ERC721',
+            },
+            collection: {
+              name: 'Collection 2574',
+              image_url: 'url',
             },
             description: 'Description 2574',
             image_url: 'image/2574.png',
